don't over generate flux kernels
